chore: moved resourceheading component into journeys ui to solve circular dependencies
