+++ libs/journeys/ui/src/components/ResourceHeading/ResourceHeading.spec.tsx
