+++ libs/journeys/ui/src/components/ResourceHeading/index.ts
